renaming to make more logical
